cicd/requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10
pydantic>=2.5.0
gitpython>=3.1.40
pyyaml>=6.0.1
jinja2>=3.1.2
requests>=2.31.0
python-dotenv>=1.0.0
llama-cpp-python>=0.2.0
transformers>=4.35.0
huggingface-hub>=0.19.0
numpy>=1.24.0
pandas>=2.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.10.0
isort>=5.12.0
mypy>=1.6.0
pylint>=3.0.0
